requirements.txt
python-dotenv==1.0.1
beautifulsoup4==4.12.3
lxml==5.2.2
lxml_html_clean==0.4.1
readability-lxml==0.8.1
requests==2.32.3
